+++ components/Instructions.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+function Instructions() {
+	return (
+		<div className="container mb-6">
+			<div className="text-lg font-bold mb-1">Instructions</div>
+      <div className="text-md">
+        <ul>
+        <li>1. Point the cursor over the row you wish to drop your piece in.</li>
+        <li>2. Left click to drop your piece</li>
+        <li>3. When you can connect four pieces vertically, horizontally or diagonally you win</li>
+        </ul>
+      </div>
+		</div>
+	);
+}
+
+export default Instructions;
