+++ package.json
@@ -13,6 +13,7 @@
 	"dependencies": {
 		"next": "^12.3.4",
 		"nextjs-progressbar": "^0.0.13",
+		"prettier": "^2.8.8",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2"
 	},
